refactor: move labs lanes and hide experimental help
