+++ openclaw_mem/labs/__init__.py
@@ -1,0 +1,9 @@
+"""Experimental memory lanes kept outside the stable core API."""
+
+__all__ = [
+    "continuity_soak",
+    "gbrain_sidecar",
+    "goal_primitive",
+    "self_improvement_surface",
+    "self_model_sidecar",
+]
